Move YamlMapConverter to file with license
