+++ lib/src/utils/extensions.dart
@@ -21,30 +21,6 @@ extension ObjectUtils<T> on T? {
   }
 }
 
-/// Source: https://github.com/epam-cross-platform-lab/swagger-dart-code-generator/blob/master/lib/src/extensions/yaml_extensions.dart
-extension YamlMapConverter on YamlMap {
-  dynamic _convertNode(dynamic v) {
-    if (v is YamlMap) {
-      return v.toMap();
-    } else if (v is YamlList) {
-      final list = <dynamic>[];
-      for (final e in v) {
-        list.add(_convertNode(e));
-      }
-      return list;
-    } else {
-      return v;
-    }
-  }
-
-  Map<String, dynamic> toMap() {
-    final map = <String, dynamic>{};
-    nodes.forEach((k, v) =>
-        map[(k as YamlScalar).value.toString()] = _convertNode(v.value));
-    return map;
-  }
-}
-
 extension LintCodeCopyWith on LintCode {
   LintCode copyWith({
     String? name,
